refactor(racingcar): Application에서 입력과 출력을 분리
