+++ src/main/kotlin/racingcar/RacingCarOutputView.kt
@@ -1,0 +1,14 @@
+package racingcar
+
+object RacingCarOutputView {
+
+    fun output(racingResults : RacingResults) {
+        for(i in racingResults.values.indices) {
+            println("라운드 - [${racingResults.values[i].round}] 결과")
+            racingResults.values[i].values.forEach() {
+                println(it)
+            }
+            println()
+        }
+    }
+}
